apps/api/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
google-genai
pydantic
python-dotenv
